wp-theme/CascadeProjects/personal-website/api_improvements.py
"""
API-integraatioiden parannukset.
"""

import os
import json
import logging
import asyncio
from typing import Dict, Any, Optional, List
from datetime import datetime
from functools import lru_cache
from cryptography.fernet import Fernet
import aiohttp
import jwt
from dataclasses import dataclass

# Konfiguroi lokitus
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s'
)
logger = logging.getLogger(__name__)

@dataclass
class APIMetrics:
    """API-metriikkojen seuranta"""
    calls: int = 0
    errors: int = 0
    latency: float = 0.0
    tokens: int = 0
    cost: float = 0.0

class SecureKeyManager:
    """Turvallinen avainten hallinta"""
    
    def __init__(self, key_file: str = ".env.key"):
        """
        Alusta avainten hallinta
        
        Args:
            key_file: Avaintiedoston polku
        """
        self.key_file = key_file
        self.fernet = None
        self._load_or_create_key()
    
    def _load_or_create_key(self):
        """Lataa tai luo salausavain"""
        if os.path.exists(self.key_file):
            with open(self.key_file, "rb") as f:
                key = f.read()
        else:
            key = Fernet.generate_key()
            with open(self.key_file, "wb") as f:
                f.write(key)
        
        self.fernet = Fernet(key)
    
    def encrypt_key(self, api_key: str) -> bytes:
        """Salaa API-avain"""
        return self.fernet.encrypt(api_key.encode())
    
    def decrypt_key(self, encrypted_key: bytes) -> str:
        """Pura API-avaimen salaus"""
        return self.fernet.decrypt(encrypted_key).decode()

class AccessManager:
    """Pääsynhallinta"""
    
    def __init__(self, secret_key: str):
        """
        Alusta pääsynhallinta
        
        Args:
            secret_key: Salainen avain
        """
        self.secret = secret_key
    
    def create_token(self, user_id: str) -> str:
        """Luo JWT token"""
        return jwt.encode(
            {"user_id": user_id},
            self.secret,
            algorithm="HS256"
        )
    
    def validate_token(self, token: str) -> Optional[str]:
        """Validoi JWT token"""
        try:
            data = jwt.decode(
                token,
                self.secret,
                algorithms=["HS256"]
            )
            return data["user_id"]
        except:
            return None

class APIOptimizer:
    """API-kutsujen optimointi"""
    
    def __init__(
        self,
        cache_size: int = 1000,
        rate_limit: int = 50,
        concurrent_limit: int = 5
    ):
        """
        Alusta optimoija
        
        Args:
            cache_size: Välimuistin koko
            rate_limit: Maksimi pyyntöä minuutissa
            concurrent_limit: Maksimi rinnakkaiset pyynnöt
        """
        self.rate_limit = rate_limit
        self.semaphore = asyncio.Semaphore(concurrent_limit)
        self.last_request = datetime.now()
        self.metrics = APIMetrics()
    
    @lru_cache(maxsize=1000)
    async def cached_request(
        self,
        func: callable,
        *args,
        **kwargs
    ) -> Any:
        """
        Välimuistitettu API-kutsu
        
        Args:
            func: API-funktio
            *args: Funktion argumentit
            **kwargs: Funktion avainsana-argumentit
        
        Returns:
            Any: API-kutsun tulos
        """
        async with self.semaphore:
            # Rate limiting
            now = datetime.now()
            elapsed = (now - self.last_request).total_seconds()
            if elapsed < 60 / self.rate_limit:
                await asyncio.sleep(60 / self.rate_limit - elapsed)
            
            start = datetime.now()
            try:
                result = await func(*args, **kwargs)
                
                # Päivitä metriikka
                self.metrics.calls += 1
                self.metrics.latency = (
                    datetime.now() - start
                ).total_seconds()
                
                return result
            
            except Exception as e:
                # Virheenkäsittely
                self.metrics.errors += 1
                logger.error(f"API error: {str(e)}")
                return None
            
            finally:
                self.last_request = datetime.now()

class BatchProcessor:
    """Eräajoprosessori"""
    
    def __init__(
        self,
        batch_size: int = 10,
        max_retries: int = 3,
        timeout: float = 30.0
    ):
        """
        Alusta prosessori
        
        Args:
            batch_size: Eräkoko
            max_retries: Maksimi uudelleenyritykset
            timeout: Timeout sekunteina
        """
        self.batch_size = batch_size
        self.max_retries = max_retries
        self.timeout = timeout
    
    async def process_batch(
        self,
        items: List[Any],
        processor: callable
    ) -> List[Any]:
        """
        Prosessoi erä
        
        Args:
            items: Prosessoitavat itemit
            processor: Prosessointifunktio
        
        Returns:
            List[Any]: Prosessoidut tulokset
        """
        results = []
        
        # Jaa eriin
        for i in range(0, len(items), self.batch_size):
            batch = items[i:i + self.batch_size]
            
            # Prosessoi erä
            batch_results = await asyncio.gather(*[
                self._process_with_retry(item, processor)
                for item in batch
            ])
            
            results.extend(batch_results)
        
        return results
    
    async def _process_with_retry(
        self,
        item: Any,
        processor: callable
    ) -> Any:
        """
        Prosessoi item uudelleenyrityksillä
        
        Args:
            item: Prosessoitava item
            processor: Prosessointifunktio
        
        Returns:
            Any: Prosessoitu tulos
        """
        for attempt in range(self.max_retries):
            try:
                async with asyncio.timeout(self.timeout):
                    return await processor(item)
            
            except asyncio.TimeoutError:
                logger.warning(
                    f"Timeout processing item: {item}"
                    f" (attempt {attempt + 1}/{self.max_retries})"
                )
            
            except Exception as e:
                logger.error(
                    f"Error processing item: {item}"
                    f" (attempt {attempt + 1}/{self.max_retries})"
                    f" - {str(e)}"
                )
            
            # Odota ennen uudelleenyritystä
            await asyncio.sleep(2 ** attempt)
        
        return None

async def main():
    """Testaa parannuksia"""
    # Turvallinen avainten hallinta
    key_manager = SecureKeyManager()
    encrypted = key_manager.encrypt_key("test-api-key")
    decrypted = key_manager.decrypt_key(encrypted)
    logger.info(f"Key test: {decrypted == 'test-api-key'}")
    
    # Pääsynhallinta
    access = AccessManager("secret")
    token = access.create_token("user123")
    user_id = access.validate_token(token)
    logger.info(f"Token test: {user_id == 'user123'}")
    
    # API optimointi
    optimizer = APIOptimizer()
    
    async def test_api():
        return "API response"
    
    result = await optimizer.cached_request(test_api)
    logger.info(f"API test: {result == 'API response'}")
    logger.info(f"API metrics: {optimizer.metrics}")
    
    # Eräajo
    processor = BatchProcessor()
    items = ["item1", "item2", "item3"]
    
    async def process_item(item: str):
        return f"Processed {item}"
    
    results = await processor.process_batch(items, process_item)
    logger.info(f"Batch results: {results}")

if __name__ == "__main__":
    asyncio.run(main())
