wp-theme/CascadeProjects/personal-website/requirements.txt
python-dotenv==1.0.0
openai==1.3.0
anthropic==0.5.0
huggingface-hub==0.19.0
cryptography>=41.0.0
aiohttp>=3.9.0
PyJWT>=2.8.0
